refactor(event cache): move relation extraction into common store helpers
